Fix file copy bug that causes local package imports to fail.
The build pack was copying the contents of $GOPATH to
$GOPATH/src/demoapp causing the resulting directory structure
to look something like:

    $GOPATH/src/demoapp/src/demoapp

This was causing local package imports to fail.  E.g. if I have
a directory "models" that defines a package models then I should be
able to import it via "demoapp/models" but the extra path addition
caused it to fail to find that directory.