Remove support for .godir and `Godeps` file
Both of these were never "supported" by heroku staff, but still
available in the buildpack for legacy reasons. They've been marked as
deprecated for about 2 montsh and the buildpack spit out warning
messages to that effect since then.

This has the nice side effect of being able to remove a bunch of extra
bash.