(WIP) GO15VENDOREXPERIMENT without godeps
After growing frustrated with `godeps`, I decided to embrace the official
vendoring solution provided by `GO15VENDOREXPERIMENT`, which is now on
by default, and will come standard in 1.7 (won't be able to disable it).

This patch is my attempt to support `GO15VENDOREXPERIMENT` in a
dependency manager agnostic way.

Personally, I am using [`glide`](https://glide.sh) to manage my `vendor/` directory,
but realized that at this juncture this buildpack doesn't need to
support any particular tool to support a more generic vendored
dependency approach, since any front-end tool is irrelevant and not used
by go when the app is compiled.

This buildpack will now detect if `$GO15VENDOREXPERIMENT` is set to
`"1"`, and that a `vendor/` directory exists. If these are found, a
simple `go install` is run (using the appropriate flags). This should
work for those without any vendored dependencies, provided they have a
placeholder file in `vendor/` to ensure the directory can be pushed to
git.

This is my first attempt at hacking on a buildpack, so any feedback is
welcome and appreciated.

A couple of loose ends that may need attention:

 - I don't test the version of go that is being used, to ensure it's 1.5 or greater.
 - The text I `warn` users about in the build step should probably link
   to articles in your devcenter, recommend that you use `godeps`, etc.