Optionally, install tools into image
If GO_INSTALL_TOOLS_IN_IMAGE is set to the string "true", then the go
toolchain and the dependency management tool is copied into the image.

The cache is still primed so we don't have to download the same go
version/dependency management tool for every build. The tools are copied
out of the cache.