Need a way to specify heroku build info
This is inspired by build tags, but I'm not sure I like this.

The idea is that it could be used like this:

// +heroku <key1> <value1> [<value2>...]

I believe all the existing options translate...

// +heroku goVersion go1.11
// +heroku install ./cmd/... ./foo
// +heroku additionalTools github.com/mattes/migrate some.other/package

I tested a bit and `go mod tidy` preserves comments, so I think the
comments would be safe from automated tooling nuking them.

This format also makes extraction super easy:

$ awk '{ if ($1 == "//" && $2 == "+heroku" && $3 == "goVersion" ) { print $4 } }' ./go.mod
go1.11

or

$ awk '{ if ($1 == "//" && $2 == "+heroku" && $3 == "install" ) { print substr($0, index($0,$4)) }}' ./go.mod
./cmd/... ./foo