Don't fetch deps when -d vendor and len(go.sum) > 0
For those that wish to have repeatable builds in the face of Github
(or other sources) outages and/or private repositories that are vendored
enforce -mod=vendor if the `vendor` directory exists AND the size of
`go.sum` is > 0.

Either don't use `vendor` or don't commit `go.sum` if you want to skip
downloading and verification of dependencies.

I suspect that once modules are taken out of their experimental phase
we'll see a decrease in people using vendoring at all and only those who
need repeatable builds (or otherwise find value in them) will opt to
continue using it.