Handle the stdlib like we handle any other file.
This makes it so that every build/test run doesn't curl the file. IME
this leads to occasional failed builds; is a PITA when in bandwidth
contrainted environments like airplanes, coffee shops; or when you
don't have a network connection.

Historically I've taken the stance that the Go Buildpack's tests
shouldn't rely on the network unless it's testing something that relies
on the network.