Update the tests to work new output & add .sh extensions
The detect package commit changes the output of the buildpack a bunch,
so this updates the tests to deal with those changes.

This was done by adding a few test helpers to utils.sh so that the
boilerplate is easier to change once vs. many times.

Also: Add .sh extensions to the shell scripts to make it clearer what
they are.