make `go list` use of -mod=vendor more dynamic
If a Go release of 1.14 or greater is used with a language version of
1.14 or greater, always passing -mod=vendor to `go list` breaks due to
new consistency checks in 1.14 (https://golang.org/doc/go1.14#vendor).

Maybe better would be to inspect the desired language version (using
`go mod edit -json | jq -r '.Go'` or so) but that could use some more
thinking.

To fix it for now, move up the adding of -mod=vendor to the build
flags and use those when calling `go list`. This should leave out
-mod=vendor if there is no vendor directory.