Add Heroku-20 to the Travis test matrix
To confirm that the buildpack works on the preview of the new stack.

The install step has also been updated to only pull the image being
tested, rather than all of the images for each job in the matrix,
saving 2 minutes per job.

I had to recompile Yara in the `godep-cgo-vendored` fixture, since it
was linked against `libssl1.0.0` which is not available for Ubuntu 20.04.

Since there isn't a single libssl version that is supported across all
of our stacks (`libssl1.1` is only available for Ubuntu 18.04+), I used
`configure --without-crypto` to remove the OpenSSL dependency entirely.

The removed files were all redundant for the purposes of that test.

Refs W-7472042.