compile: set GOPATH earlier for modules
In https://github.com/heroku/heroku-buildpack-go/pull/396 the explicit
`-mod=vendor` was removed from the `go list` command used to find main
packages to install. This enabled it to fetch modules as necessary for
listing if vendoring was not being used. When modules are fetched,
they are fetched into $GOPATH/pkg/mod. If GOPATH is not set, it
defaults to $HOME/go.

In build mode, at least Heroku runs compile with a build directory
that is not the same as $HOME. In test mode, however, at least Heroku
CI runs test-compile with a build directory that is equal to
$HOME.

When this initial `go list` was being run, GOPATH was not yet set,
which in test mode downloaded modules to $HOME/go. Later, once `go
test ./...` or similar was run in $HOME, the source of the downloaded
modules was being considered. This caused errors such as:

go: directory go/pkg/mod/github.com/gorilla/mux@v1.6.2 outside available modules

To fix this, set GOPATH for modules earlier, before `go list` is run.

To get better test fidelity, run testpack tests with HOME set to the
temporary build directory.