Switch to Circle CI (#461)
Since we're migrating all repos away from Travis CI.

Previously the tests were run via `make test`, which used to:
- Copy parts (but not all) of the buildpack to a temporary location,
  to try and emulate the packaged version of the buildpack (though
  the handling was out of sync with `buildpack.toml`, so would not
  protect against `buildpack.Ignore` related bugs)
- Invoke the tests using `docker run`, with that temporary
  location bind mounted into the container

On Circle CI by default jobs run in a container that does not have
access to a Docker daemon. Access to a remote daemon can be enabled via
the `setup_remote_docker` feature, however since it's a remote daemon,
it does not support bind mounting local directories.

As such the options are to either:
1. Run Circle CI jobs using the `machine` executor (full VM), which has
  slower startup times
2. Switch from `docker run` to a `docker build` approach to running the
  tests, but this would impact test iteration times locally
3. Using `docker cp` to emulate the bind mount, however that would
  mean copying the large test assets to the remote machine, and means
  CI couldn't share the same make target as local development anyway
4. Use native Circle container support, and in CI run the tests directly
  in the circle environment.

This change implements (4), since it's the most idiomatic way of running
tests on Circle CI.

The `sbin/copy` file has been removed, since:
- it doesn't work locally on macOS (presume `tar` differences)
- its approach hasn't aged well, now that the files included in the
  published buildpack are controlled via `buildpack.Ignore`
- copying files out of the project directory is unnecessary both in
  CI and when running tests locally in docker

The secrets in the old `.travis.yml` are for running the private
dependency tests, however there is no documentation of what account
owns these, how to rotate them etc. As such I've omitted them for now,
since the couple of tests that use them are automatically skipped if
they are not set. They can be added via the Circle CI dashboard if
needed later (though I'm not a fan of having tests rely on fragile
credentials that expire/be offboarded at some point in the future).

Closes GUS-W-9115160.