Add Heroku-22 to the CI test matrix
The Go binaries are not stack-specific, so no code/binary changes
are needed for Heroku-22 - only testing it in CI.

This buildpack uses Docker tests rather than Hatchet tests, so the
new stack can be added to the CI test matrix even before it's
possible to create new Heroku-22 apps on the platform.

GUS-W-10344045.