Go 1.25.0, bump default to 1.24, and a few smaller fixes (#600)
* Fix incorrect changelog for v210/1.25rc3

The mapping for 1.25 to 1.25rc3 was not in that release

Refs #598, #599

GUS-W-19293798

* Fix missing version bumps in test assets for v210

Forgotten in #598

GUS-W-19293798

* Add missing tests for go1.22, go1.23, go1.24

* Drop redundant version expansions

We only need the two-group case, for mapping to the latest point releases.

The old 1.20.0 -> 1.20, 1.19.0 -> 1.19 etc mappings were necessary because before version 1.21.0, there was never a .0 release, hence that alias so people could explicitly opt into the first release of a series, while the "shorthand" (e.g. 1.20) always mapped to the latest.

Not necessary anymore since the versioning scheme changed with go1.21.0:

$ curl -sI https://dl.google.com/go/go1.20.linux-amd64.tar.gz | head -n1
HTTP/2 200
$ curl -sI https://dl.google.com/go/go1.20.0.linux-amd64.tar.gz | head -n1
HTTP/2 404

$ curl -sI https://dl.google.com/go/go1.21.linux-amd64.tar.gz | head -n1
HTTP/2 404
$ curl -sI https://dl.google.com/go/go1.21.0.linux-amd64.tar.gz | head -n1
HTTP/2 200

GUS-W-19331843

* Add go1.25.0

GUS-W-19331787

* Bump default version to go1.24

GUS-W-19331849